requirements.txt
flask==3.0.0
google-generativeai==0.3.2
yt-dlp==2023.12.30
whisper==1.1.10
TTS==0.22.0
pydub==0.25.1
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
